+++ src/components/AdminCheck.js
@@ -1,0 +1,17 @@
+const showAdmin = () => {
+  if (isAdmin) {
+    return (
+      //stuff
+      null
+    );
+  }
+};
+
+const checkAdmin = () => {
+  //Get datas and if admin then set state isAdmin = true
+};
+
+const login = () => {
+  //Fetch/Axios to server
+  //Return the cookie?
+};
